Move Brushed ESC auto detection into seperate source files pwm_output.c and pwm_output_hal.c are used aternativly
