+++ src/main/drivers/pwm_output.h
@@ -114,15 +114,3 @@ pwmOutputPort_t *pwmGetMotors(void);
 bool pwmIsSynced(void);
 void pwmDisableMotors(void);
 void pwmEnableMotors(void);
-
-#ifdef BRUSHED_ESC_AUTODETECT
-typedef enum {
-    MOTOR_UNKNOWN = 0,
-    MOTOR_BRUSHED,
-    MOTOR_BRUSHLESS
-} HardwareMotorTypes_e;
-
-extern uint8_t hardwareMotorType;
-
-void detectBrushedESC(void);
-#endif
